requirements.txt
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
mpmath==1.3.0
numpy==1.24.3
onnxruntime-gpu==1.14.1
opencv-python==4.5.4.60
packaging==23.1
protobuf==4.23.2
sympy==1.12
kafka-python
scipy
boto3
imutils
python-dotenv
onnxruntime-gpu
dlib
